algo/src/test/java/com/hartwig/actin/algo/evaluation/treatment/HasHadAnyCancerTreatmentTest.java
package com.hartwig.actin.algo.evaluation.treatment;

import static com.hartwig.actin.algo.evaluation.EvaluationAssert.assertEvaluation;

import java.util.List;

import com.google.common.collect.Lists;
import com.hartwig.actin.algo.datamodel.EvaluationResult;
import com.hartwig.actin.clinical.datamodel.PriorTumorTreatment;

import org.junit.Test;

public class HasHadAnyCancerTreatmentTest {

    @Test
    public void canEvaluate() {
        HasHadAnyCancerTreatment function = new HasHadAnyCancerTreatment();

        // No treatments
        List<PriorTumorTreatment> treatments = Lists.newArrayList();
        assertEvaluation(EvaluationResult.FAIL, function.evaluate(TreatmentTestFactory.withPriorTumorTreatments(treatments)));

        // One random treatment
        treatments.add(TreatmentTestFactory.builder().build());
        assertEvaluation(EvaluationResult.PASS, function.evaluate(TreatmentTestFactory.withPriorTumorTreatments(treatments)));
    }
}